fixera build igen

+++ src/server/game/Entities/Player/Player.cpp
@@ -12186,7 +12186,7 @@ PartyResult Player::CanUninviteFromGroup(ObjectGuid targetPlayerGUID) const
             if (Aura* dungeonCooldownAura = target->GetAura(lfg::LFG_SPELL_DUNGEON_COOLDOWN))
             {
                 int32 elapsedTime = dungeonCooldownAura->GetMaxDuration() - dungeonCooldownAura->GetDuration();
-                if (sWorld->getIntConfig(CONFIG_LFG_KICK_PREVENTION_TIMER) > elapsedTime)
+                if (static_cast<int32>(sWorld->getIntConfig(CONFIG_LFG_KICK_PREVENTION_TIMER)) > elapsedTime)
                 {
                     return ERR_PARTY_LFG_BOOT_NOT_ELIGIBLE_S;
                 }
