+++ src/server/game/Handlers/GroupHandler.cpp
@@ -341,7 +341,7 @@ void WorldSession::HandleGroupUninviteGuidOpcode(WorldPacket& recvData)
                 if (Aura* dungeonCooldownAura = kickTarget->GetAura(lfg::LFG_SPELL_DUNGEON_COOLDOWN))
                 {
                     int32 elapsedTime = dungeonCooldownAura->GetMaxDuration() - dungeonCooldownAura->GetDuration();
-                    if (sWorld->getIntConfig(CONFIG_LFG_KICK_PREVENTION_TIMER) > elapsedTime)
+                    if (static_cast<int32>(sWorld->getIntConfig(CONFIG_LFG_KICK_PREVENTION_TIMER)) > elapsedTime)
                     {
                         SendPartyResult(PARTY_OP_UNINVITE, name, res, (sWorld->getIntConfig(CONFIG_LFG_KICK_PREVENTION_TIMER) - elapsedTime) / 1000);
                     }
